new phase management
