+++ pandar_pointcloud/include/pandar_pointcloud/decoder/pandar40_decoder.hpp
@@ -31,7 +31,13 @@ public:
     DUAL_ONLY,
   };
 
-  Pandar40Decoder(rclcpp::Node & node, Calibration& calibration, float scan_phase = 0.0f, double dual_return_distance_threshold = 0.1, ReturnMode return_mode = ReturnMode::DUAL);
+  Pandar40Decoder(rclcpp::Node &node,
+                  Calibration &calibration,
+                  double scan_phase,
+                  const std::vector<double> &angle_range,
+                  const std::vector<double> &distance_range,
+                  double dual_return_distance_threshold = 0.1,
+                  ReturnMode return_mode = ReturnMode::DUAL);
   void unpack(const pandar_msgs::msg::PandarPacket& raw_packet) override;
   bool hasScanned() override;
   PointcloudXYZIRADT getPointcloud() override;
@@ -54,6 +60,8 @@ private:
 
   std::array<size_t, LASER_COUNT> firing_order_;
 
+  std::vector<int> angle_range_;
+  std::vector<double> distance_range_;
   ReturnMode return_mode_;
   double dual_return_distance_threshold_;
   Packet packet_;
@@ -61,9 +69,11 @@ private:
   PointcloudXYZIRADT scan_pc_;
   PointcloudXYZIRADT overflow_pc_;
 
-  uint16_t scan_phase_;
+  int scan_phase_;
   int last_phase_;
   bool has_scanned_;
+  bool use_overflow_;
+  bool reset_scan_;
 };
 
 }  // namespace pandar40
